\name{equateIRT-internal}
\alias{approxabs}
\alias{aug_lagr_pen}
\alias{der_aug_lagr_pen}
\alias{derfpen}
\alias{derridgepennominal}
\alias{fpen}
\alias{gradnominallik}
\alias{gradnominallik_fpen}
\alias{gradnominallikaug}
\alias{gradnominallikaugRcppA}
\alias{gradnominallik_fpenRcppA}
\alias{gradnominallikRcppA}
\alias{nominal_proxgr}
\alias{nominallik}
\alias{nominallik_fpen}
\alias{nominallikaug}
\alias{nominallikaugRcppA}
\alias{nomprobs}
\alias{nominal_grfused}
\alias{nominallik_fpenRcppA}
\alias{nominallikRcppA}
\alias{reduce_data}
\alias{par2list}
\alias{ridgepennominal}
\alias{simdatanom}
\alias{sim_reg_nom}
\alias{vinit}
\alias{w2W}

\title{Internal Functions of Package regIRT}
\description{
  Internal Functions of Package regIRT.
}
\details{
  These are not to be called by the user.
}
\keyword{ internal }
